+++ sdk/resourcemanager/cosmos/armcosmos/percentilesourcetarget_client.go
@@ -46,7 +46,7 @@ func NewPercentileSourceTargetClient(subscriptionID string, credential azcore.To
 // NewListMetricsPager - Retrieves the metrics determined by the given filter for the given account, source and target region.
 // This url is only for PBS and Replication Latency data
 //
-// Generated from API version 2023-11-15-preview
+// Generated from API version 2024-02-15-preview
 //   - resourceGroupName - The name of the resource group. The name is case insensitive.
 //   - accountName - Cosmos DB database account name.
 //   - sourceRegion - Source region from which data is written. Cosmos DB region, with spaces between words and each word capitalized.
@@ -108,8 +108,8 @@ func (client *PercentileSourceTargetClient) listMetricsCreateRequest(ctx context
 		return nil, err
 	}
 	reqQP := req.Raw().URL.Query()
-	reqQP.Set("api-version", "2023-11-15-preview")
 	reqQP.Set("$filter", filter)
+	reqQP.Set("api-version", "2024-02-15-preview")
 	req.Raw().URL.RawQuery = reqQP.Encode()
 	req.Raw().Header["Accept"] = []string{"application/json"}
 	return req, nil