+++ sdk/resourcemanager/cosmos/armcosmos/restorablesqlcontainers_client.go
@@ -47,7 +47,7 @@ func NewRestorableSQLContainersClient(subscriptionID string, credential azcore.T
 // This helps in scenario where container was accidentally deleted. This API requires
 // 'Microsoft.DocumentDB/locations/restorableDatabaseAccounts/…/read' permission
 //
-// Generated from API version 2023-11-15-preview
+// Generated from API version 2024-02-15-preview
 //   - location - Cosmos DB region, with spaces between words and each word capitalized.
 //   - instanceID - The instanceId GUID of a restorable database account.
 //   - options - RestorableSQLContainersClientListOptions contains the optional parameters for the RestorableSQLContainersClient.NewListPager
@@ -96,15 +96,15 @@ func (client *RestorableSQLContainersClient) listCreateRequest(ctx context.Conte
 		return nil, err
 	}
 	reqQP := req.Raw().URL.Query()
-	reqQP.Set("api-version", "2023-11-15-preview")
+	reqQP.Set("api-version", "2024-02-15-preview")
+	if options != nil && options.EndTime != nil {
+		reqQP.Set("endTime", *options.EndTime)
+	}
 	if options != nil && options.RestorableSQLDatabaseRid != nil {
 		reqQP.Set("restorableSqlDatabaseRid", *options.RestorableSQLDatabaseRid)
 	}
 	if options != nil && options.StartTime != nil {
 		reqQP.Set("startTime", *options.StartTime)
-	}
-	if options != nil && options.EndTime != nil {
-		reqQP.Set("endTime", *options.EndTime)
 	}
 	req.Raw().URL.RawQuery = reqQP.Encode()
 	req.Raw().Header["Accept"] = []string{"application/json"}