+++ sdk/resourcemanager/cosmos/armcosmos/throughputpools_client.go
@@ -45,7 +45,7 @@ func NewThroughputPoolsClient(subscriptionID string, credential azcore.TokenCred
 
 // NewListPager - Lists all the Azure Cosmos DB Throughput Pools available under the subscription.
 //
-// Generated from API version 2023-11-15-preview
+// Generated from API version 2024-02-15-preview
 //   - options - ThroughputPoolsClientListOptions contains the optional parameters for the ThroughputPoolsClient.NewListPager
 //     method.
 func (client *ThroughputPoolsClient) NewListPager(options *ThroughputPoolsClientListOptions) *runtime.Pager[ThroughputPoolsClientListResponse] {
@@ -83,7 +83,7 @@ func (client *ThroughputPoolsClient) listCreateRequest(ctx context.Context, opti
 		return nil, err
 	}
 	reqQP := req.Raw().URL.Query()
-	reqQP.Set("api-version", "2023-11-15-preview")
+	reqQP.Set("api-version", "2024-02-15-preview")
 	req.Raw().URL.RawQuery = reqQP.Encode()
 	req.Raw().Header["Accept"] = []string{"application/json"}
 	return req, nil
@@ -100,7 +100,7 @@ func (client *ThroughputPoolsClient) listHandleResponse(resp *http.Response) (Th
 
 // NewListByResourceGroupPager - List all the ThroughputPools in a given resource group.
 //
-// Generated from API version 2023-11-15-preview
+// Generated from API version 2024-02-15-preview
 //   - resourceGroupName - The name of the resource group. The name is case insensitive.
 //   - options - ThroughputPoolsClientListByResourceGroupOptions contains the optional parameters for the ThroughputPoolsClient.NewListByResourceGroupPager
 //     method.
@@ -143,7 +143,7 @@ func (client *ThroughputPoolsClient) listByResourceGroupCreateRequest(ctx contex
 		return nil, err
 	}
 	reqQP := req.Raw().URL.Query()
-	reqQP.Set("api-version", "2023-11-15-preview")
+	reqQP.Set("api-version", "2024-02-15-preview")
 	req.Raw().URL.RawQuery = reqQP.Encode()
 	req.Raw().Header["Accept"] = []string{"application/json"}
 	return req, nil